+++ bdd-cucumber-gherkin-lib-security/src/main/kotlin/com/ragin/bdd/cucumber/security/models/ProxyEndpoint.kt
@@ -1,0 +1,7 @@
+package com.ragin.bdd.cucumber.security.models
+
+/** Host and port of the HTTP proxy a security scanner listens on. */
+data class ProxyEndpoint(
+    val host: String,
+    val port: Int
+)